examples/steamdeck/pyproject.toml
[project]
name = "steamdeck-controller"
version = "0.1.0"
description = "Steam Deck controller WebSocket bridge and test utilities"
requires-python = ">=3.12"

dependencies = [
    "evdev-binary",
    "aiohttp",
]

[project.optional-dependencies]
test = [
    "websockets",
    "httpx",
]
